Move modals into their own folder. Persist the changes in other pages where they are used and update the solution proj fiile.
