+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/BridelleBelleMobileApplication.projitems
@@ -25,24 +25,24 @@
     <Compile Include="$(MSBuildThisFileDirectory)Models\RegisteredUser.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Types\SignInRegisterResponse.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ViewModels\MagazineViewModel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Views\CheckoutOptions.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Views\Modals\CheckoutOptions.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Views\MagazineInformation.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Views\MagazinePreview.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Views\Modals\MagazinePreview.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Views\MagazineViewingOptions.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Views\Modals\MagazineViewingOptions.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Views\RegisterUser.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Views\Modals\RegisterUser.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Models\User.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ViewModels\UserViewModel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Views\AdvertiserChoicePopup.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Views\Modals\AdvertiserChoicePopup.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Views\AdvertiserInformation.xaml.cs">
@@ -95,7 +95,7 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\AdvertiserChoicePopup.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\Modals\AdvertiserChoicePopup.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
@@ -115,12 +115,12 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\RegisterUser.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\Modals\RegisterUser.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\MagazinePreview.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\Modals\MagazinePreview.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
@@ -130,13 +130,16 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\MagazineViewingOptions.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\Modals\MagazineViewingOptions.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\CheckoutOptions.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Views\Modals\CheckoutOptions.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="$(MSBuildThisFileDirectory)Views\Modals\" />
   </ItemGroup>
 </Project>
