+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/Views/Modals/MagazinePreview.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <pages:PopupPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="BridelleBelleMobileApplication.Views.MagazinePreview"
+             x:Class="BridelleBelleMobileApplication.Views.Modals.MagazinePreview"
             xmlns:forms="clr-namespace:Xamarin.Forms;assembly=Xamarin.Forms.CarouselView"
             xmlns:pages="clr-namespace:Rg.Plugins.Popup.Pages;assembly=Rg.Plugins.Popup"
 >
