+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/Views/Modals/MagazineViewingOptions.xaml
@@ -2,19 +2,16 @@
 <pages:PopupPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:pages="clr-namespace:Rg.Plugins.Popup.Pages;assembly=Rg.Plugins.Popup"
-                 x:Class="BridelleBelleMobileApplication.Views.AdvertiserChoicePopup">
-
+             x:Class="BridelleBelleMobileApplication.Views.Modals.MagazineViewingOptions">
     <StackLayout VerticalOptions="Center" HorizontalOptions="FillAndExpand" Padding="20, 20, 20, 20">
         <StackLayout Padding="0, 10, 0, 0" BackgroundColor="#F4EDEC">
-            <Label Text="Choose an advertiser" TextColor="Black" FontSize="20" HorizontalOptions="Center"></Label>
             <ScrollView>
                 <StackLayout>
-                    <StackLayout Orientation="Horizontal">
+                    <StackLayout Orientation="Horizontal" HeightRequest="150">
                         <ListView x:Name="options" ItemSelected="OpenAdvertiser"/>
                     </StackLayout>
                 </StackLayout>
             </ScrollView>
         </StackLayout>
     </StackLayout>
-
 </pages:PopupPage>
