+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/Views/Modals/RegisterUser.xaml
@@ -2,7 +2,7 @@
 <pages:PopupPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:pages="clr-namespace:Rg.Plugins.Popup.Pages;assembly=Rg.Plugins.Popup"
-             x:Class="BridelleBelleMobileApplication.Views.RegisterUser">
+             x:Class="BridelleBelleMobileApplication.Views.Modals.RegisterUser">
     <StackLayout VerticalOptions="Center" HorizontalOptions="FillAndExpand" Padding="20, 20, 20, 20">
         <StackLayout Padding="0, 10, 0, 0" BackgroundColor="#F4EDEC">
             <Label Text="Register" TextColor="Black" FontSize="20" HorizontalOptions="Center"></Label>
