+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/Views/Modals/RegisterUser.xaml.cs
@@ -10,7 +10,7 @@ using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 using BridelleBelleMobileApplication.Helpers;
 using BridelleBelleMobileApplication.Types;
-namespace BridelleBelleMobileApplication.Views
+namespace BridelleBelleMobileApplication.Views.Modals
 {
 	[XamlCompilation(XamlCompilationOptions.Compile)]
 	public partial class RegisterUser : PopupPage
